MGDOBR-1018: auto creation of global error handler topic at manager startup
