+++ manager/src/main/resources/application.properties
@@ -168,6 +168,5 @@ kafka.bootstrap.servers=${event-bridge.kafka.bootstrap.servers}
 kafka.sasl.mechanism=PLAIN
 kafka.security.protocol=SASL_SSL
 kafka.sasl.jaas.config=org.apache.kafka.common.security.plain.PlainLoginModule required username="${event-bridge.kafka.client.id}" password="${event-bridge.kafka.client.secret}";
-mp.messaging.incoming.errors.connector=smallrye-kafka
-mp.messaging.incoming.errors.topic=${event-bridge.resource-prefix}err
-mp.messaging.incoming.errors.value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
+mp.messaging.incoming.error-handler.connector=smallrye-kafka
+mp.messaging.incoming.error-handler.value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
